Refactor ckeditor config
